backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
scikit-learn==1.4.0
pydantic==2.5.3
python-dotenv==1.0.0
